Created Fair entity! WIP
